# feat: optimizations to attention.py