Taken from https://github.com/Birch-san/stable-diffusion/commit/27e645af1aa45d46ab7a984ed77e9f238efd6fcb which in turn is itself a cherry-pick of https://github.com/lstein/stable-diffusion/pull/540

This I believe will allow to finally incorporate the changes offered in #771

My small rundown in normal mode with the classic k_euler/20 steps:
|                | Dev                                            | Pull Request                                   |
|----------------|------------------------------------------------|------------------------------------------------|
| Batch_size 10  | (6.67s per image) Peak memory usage: 10326 MiB | (5.9s per image) Peak memory usage: 9485 MiB   |
| Batch_count 10 | (8.4s per image) Peak memory usage: 7342 MiB   | (7.5s per image) Peak memory usage: 6500 MiB   |
| Batch_size 15  | OOM                                            | (5.12s per image) Peak memory usage: 8819 MiB  |
| Batch_size 20  | OOM                                            | (5.22s per image) Peak memory usage: 11257 MiB |

So at least in my case it provides both memory and speed improvements with zero drawbacks. 🚀

Original message below:
this is a collaboration incorporating a lot of people's contributions -- including for example @Doggettx and the original code from @neonsecret on which the Doggetx optimizations were based (see https://github.com/lstein/stable-diffusion/issues/431, https://github.com/sd-webui/stable-diffusion-webui/pull/771\#issuecomment-1239716055). Takes exactly the same amount of time to run 8 steps as original CompVis code does (10.4 secs, ~1.25s/it).